feat(data-table): better organization
